quit game at any time
do you want to add this quit at anytime thing? it's only about 4 extra lines of code. 
i'm not saying that the player MUST quit in the middle of the game, but it's just like a safety blanket we can use if veena questions the validity of the "press quit to exit" thing. and i promise it won't make the terminal look messy. 
idk your choice. but it can't hurt ig